implemented better node exception handling - now run() raises an exception with all necessary failed node information (input fields, output fields, attributes, traceback); implemented convenience print_exception for stream runtime error; reflect change in tests; added SQL database target node; updated documentation + added reference target markup to generated node reference

+++ brewery/ds/sql_streams.py
@@ -29,7 +29,7 @@ try:
         "date": sqlalchemy.types.Date,
         "time": sqlalchemy.types.DateTime,
         "integer": sqlalchemy.types.Integer,
-        "numeric": sqlalchemy.types.Numeric,
+        "float": sqlalchemy.types.Numeric,
         "boolean": sqlalchemy.types.SmallInteger
     }
 except:
